MH-3051 - first version of FileCopy utility that does concurrent, buffered file system copies.  Requires source and destination directories with set of files in the source.  typical usage -
java -cp matterhorn-test-harness-0.8-SNAPSHOT-jar-with-dependencies.jar org.opencastproject.integrationtest.FileCopyTests -s /Users/john/dev/source -d /Users/john/dev/destination -n 2  

where n is the number of concurrent copies to be made.  Note that this number must be <= the number of files in the source directory and the directory paths must be full paths in this version.



git-svn-id: https://opencast.jira.com/svn/MH/trunk@6896 e6edf6fb-f266-4316-afb4-e53d95876a76